Banners are now served through the REST API and the main landing URL page provides items and banners information. The rest of the home endpoints such as the search query and the one-item page still need to be implemented to fit RESTful service.

+++ corner-food-market-website/src/main/java/com/cornerfoodmarketwebsite/data/single_table/entity/Discount.java
@@ -9,7 +9,7 @@ public class Discount {
     @Id
     @Column(name = "ID")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private short id;
+    private int id;
     @Column(name = "ITEM_ID")
     private short itemId;
 
@@ -56,7 +56,7 @@ public class Discount {
     @Column(name = "MODIFIED_AT")
     private Date modifiedAt;
 
-    public short getId() {
+    public int getId() {
         return id;
     }
